스크롤 동기화를 줄바꿈/코드접기에 무관하게 수정
직전 구현은 "원본 줄 간격과 렌더 픽셀 간격이 비례한다" 고 가정했다. 이 가정은
틀렸다. 에디터에서 접히지도 줄바꿈되지도 않은 한 줄이, 좁은 프리뷰 창에서는
여러 줄로 접혀 몇 배 높이를 차지한다. 반대로 여러 원본 줄이 한 문단으로 합쳐져
한 줄 높이만 차지하기도 한다. 양쪽 다 실제 렌더 결과를 재서 매핑하도록 바꿨다.

에디터 쪽 (전부 픽셀 기준으로 전환)
- lineAtViewportRatio/scrollLineToViewportRatio 를
  fractionalLineAtViewportRatio/scrollFractionalLineToViewportRatio 로 교체.
- 읽기: SCI_POSITIONFROMPOINT 로 "그 높이에 실제로 보이는 문자" 를 얻고,
  SCI_WRAPCOUNT 로 그 줄 안에서의 위치까지 소수로 표현한다. 12.5 는 12번 줄의
  세로 중간이라는 뜻이다.
- 쓰기: SCI_POINTYFROMPOSITION 으로 현재 픽셀 위치를 재고 목표와의 차이만큼
  SCI_LINESCROLL 로 화면 행 단위 스크롤한다. 문서 줄 번호로 계산하던 기존
  방식(SCI_VISIBLEFROMDOCLINE + SETFIRSTVISIBLELINE)은 자동 줄바꿈이 켜져 있으면
  어긋난다.

프리뷰 쪽 (요소 내부 비례 -> 기준점 구간 보간)
- anchorTable(): (src, 시작줄) 마다 가장 작은 요소의 실제 top 좌표를 모아
  y 순으로 정렬하고 단조성을 강제한다.
- documentYForLine / lineForDocumentY 는 이제 인접 기준점 **쌍 사이를** 선형
  보간한다. 두 기준점 사이의 픽셀 거리에는 프리뷰 쪽 자동 줄바꿈이 이미
  반영돼 있으므로, 구간마다 접힘 정도가 달라도 맞는다.
- 창 너비가 바뀌면 줄바꿈 위치가 달라지므로 기준점 캐시를 무효화한다.
- include 경계(다른 src)에서는 보간하지 않고 가까운 쪽으로 붙인다.
- 소수 줄을 그대로 주고받는다 (브리지 시그니처도 double 로).

검증: 소스 줄 간격은 전부 2줄로 같지만 렌더 높이는 크게 다른 문서
(짧은 문단 / 긴 문단 교대)를 폭 310px 프리뷰에서 측정.
  L4->L6  dLine=2 dY=41
  L6->L8  dLine=2 dY=445
  L8->L10 dLine=2 dY=41
  L10->L12 dLine=2 dY=445
같은 2줄 간격이 41px 과 445px 로 약 11배 차이 난다. 줄 수 비례로 계산했다면
네 번째 문단쯤에서 1000px 이상 어긋난다. 왕복 변환은 checked=10 bad=0.

+++ src/core/solPreviewBridge.cpp
@@ -76,7 +76,7 @@ void PreviewBridge::resetReady()
     ready_ = false;
 }
 
-void PreviewBridge::requestScrollToLine( const int sourceIndex, const int line, const double ratio )
+void PreviewBridge::requestScrollToLine( const int sourceIndex, const double line, const double ratio )
 {
     if( !ready_ )
     {
@@ -116,12 +116,12 @@ void PreviewBridge::ready( int /*protocolVersion*/ )
     }
 }
 
-void PreviewBridge::sourceLocationClicked( const int sourceIndex, const int line, const double viewportRatio )
+void PreviewBridge::sourceLocationClicked( const int sourceIndex, const double line, const double viewportRatio )
 {
     emit editorNavigationRequested( sourceIndex, line, viewportRatio );
 }
 
-void PreviewBridge::previewScrolled( const int sourceIndex, const int line, const double viewportRatio )
+void PreviewBridge::previewScrolled( const int sourceIndex, const double line, const double viewportRatio )
 {
     emit previewScrollChanged( sourceIndex, line, viewportRatio );
 }
